"Fixed" unorderable types issue on Python3
In Python 3.x, if you attempt to compare to intrinsically unorderable
types, like an integer and a float, it will return a TypeError.  This is
in contrast to Python 2.x, where it will sort them based on the name
of their object types (i.e. ints before strs).  To make natsort a bit
more robust for users who expect it to "just work", it will append
an empty string to the front of the parsed tuples if the first element
is not a string.  This will group all those non-strings together and
caused them to be sorted amongst themselves, which works around the
problem.