"Fixed" unorderable types error (again)
There is a corner case where "signed=True" and some strings have two
numbers that are adjacent to each other (such as ["5+5", "5-5", "5*5"])
where an unorderable types error can be occur.  The previous fix was for
an unorderable types at the beginning of the string; this is for the
middle of the string.

This was fixed by putting a blank string between two adjacent numbers in
the list parsed by natsort_key.  Because this is fairly expensive, and
only a small number of input data will have this problem, the sorting
is attempted without this fix, and if the unorderable types TypeError
is raised, the sorting is re-attempted with this fix turned on.