Fixed Python 2.6 set literal bug.
The set literal notation ({...}) was not introduced until Python 2.7, so
I changed the set creation to set([...]).  This set constant creation
was moved outside the function for performance considerations.

Also added a coverage badge to the README, and removed pypy from the
TravisCI config file (I was getting segfaults with py.test).