Adopted new fastnumbers module.
This module quickly confers strings to numbers, and will speed up
the natsort algorithm significantly (if it is installed). In case it is
not installed, a "fake_fastnumbers" module is added to natsort that
simulates the needed functions.

It is found that up to 30% gain in efficiency can be had by using
the fastnumbers module.