Modified API scheme and added case-based sorting.
The natsort API has been modified to be based on the value of an input enum,
rather than be based on boolean values of multiple different keyword arguments.
This will make extending easier in the future, as well as makes maintainance of
the documentation simpler.

In doing this, it made adding modifying how the sorting algorithm interprets
the case much easier.  natsort now understands case-insensitive sorting, as
well as putting lowercase first, or grouping upper and lowercases together.
Also, locale-aware sorting is now supported.

Additionally, private variables in natsort now have a leading "_".