Completed unittests. Fixed locale bug.
Added unit tests to test all the new features.

Fixed a bug where locale-aware sorting was not doing anything.  It
was found that this was due to the locale module itself... a bit of
Googling found that on Mac there are some known bugs.  To remedy this,
natsort will use PyICU if it is installed to do the sorting.

Documentation has been updated to account for the new locale-aware
sorting.

This also resolves #14.