Fixed TypeError when sorting numbers with LOCALE.
The fix for issue #7 (preventing the "unorderable types" TypeError
between numbers and strings on Python 3) can raise an "unorderable
types" TypeError between str and bytes on Python 3 if LOCALE is enabled.
This is because the strxfrm function converts str to bytes when
performing the string collation. This is because the empty string that
is inserted for numbers is of the str type but did not get sent to
strxfrm and thus did not get converted to bytes.

The fix was to add either an empty bytes string or str string depending
on if LOCALE is enabled or not.

This fixes #22.