Modified default sorting algorithm of natsort.
The old default sorting algorithm was signed floats. The new default
is unsigned integers. This aligns with the most common use case which
is to sort based on digits or version numbers, and thus will meet the
principle of least astonishment for new users.

The old versorted function is now identical to the natsorted. There is
now a realsorted function which matches the old behavior of natsorted.

This resolves #20.