Fixed bugs related to the locale library.
On some systems (particularly BSD) the locale library is hopelessly
broken. It turns out that replacing the UTF-8 locale with ISO8859-1
mitigates the issue of sorting characters with accents properly, but
it does not solve the ordering of capital and lowercase letters.

A method to detect if the locale library cannot sort these properly
and then respond to this has been added to natsort. The intent is that
natsort will handle the locale shortcomings without the user needing
to install PyICU.