Updated configuration files.
Added all dependencies to the tox.ini, and removed python3.2.

Removed 3.2 from the travis config, and added nightly builds and pypy.

The hypothesis version mentioned in the setup.py was so old that it
confused pip; this hypothesis version specification was removed.

In both setup.py and travis config, the doctests have been consolidated.