The _path_splitter helper function now uses deque.
Previously, the path components were split apart and created in reverse
order, then reversed at the end of the operations.  Now the components
are added in the correct order because of deque's fast appendleft
method.

Additionally, the imported os.path functions were renamed to make
their origin a bit clearer.