Improved fake_fastnumbers performance.
Rather than use regular expressions, we can rely on the fact that
the strings being passed to these functions will have been split
in such a way that they are either valid numbers or they are not.
So, we check if the fist digit character looks valid, and if so try
to convert. Otherwise do not.