_num_splitter is now iterable-based (eliminating TYPESAFE!).
The _num_splitter function now uses iterables/generators and function
chaining. This will reduce the overall number of loops through the
results.

A side benefit is it is now cheap to perform what the _py3_safe function
was doing, so this is done always by default. The old TYPESAFE option
to ns is now a no-op.