Redefined _natsort_key in a functional style.
This required changing the input to be the functions created by
all the function factories that have been recently committed. This
had a cascading effect that required a re-write of several unit tests,
as well as the top level natsort_keygen function, which now has the
job of creating the appropriate functions on the fly to pass to
_natsort_key.

A future commit will add more testing to natsort_keygen.

For the moment, handling of locale-specific numbers is disabled, but
will be added back in a future commit.