Significantly improved clarity of examples.
The examples in the documentation used to be rather contrived. They
are now updated to be more "real-world".

Additionally, I have made attempts to make the purpose and use
of natsort as clear as possible in the Quick Start section based
on the nature of questions that are being asked about natsort on
StackOverflow.