Move input normalization to an earlier stage.
The input normalization has been moved out of the "input_transform"
function (which was called in the "parse_string" function) and now
is the first step of the "parse_string" function. This is because
the data needs to be normalized even if the "input_transform" function
is skipped.

Tests have been reworked to understand this change.