Update hypothesis tests to not rely on "assume".
Rather than generating testing values and then removing them in
post-processing, now test values are created correct-by-construction
using the newer .filter API of hypothesis strategies.

Whilst doing this filtering, it was found that the range_check
function was doing the unnecessary step of converting its input to
floats, so this has been removed (which simplifies testing).

Last, some of the tests for the command-line --filter arguments have
been simplified because the ROI of the thorough hypothesis tests
was low.