Fix OSX-related bugs in cmp testing.
Because OSX is a "dumb sorting" OS when locale is turned on, the
method by which "dumb sorting" was being tested for natcmp worked
on non-OSX OSs, but failed on OSX. This is addressed by using patch
for both True and False returns of dumb_sort.

For the test of using natcmp with an alg, if hypothesis yields very
extreme integers then conversion of those integers into a float
(due to alg.REAL) can cause the sorting to give unexpected results.
Switching to alg.SIGNED fixes this.