src/natsort -> natsort/
The decision to move to src/natsort was motivated by the desire to
force testing on what is actually installed, not on what is currently
being developed - users would be required to use virtual environments.

For users that did not do this already it was unexpected at the least.

It also turns out that by default, tox tests against the installed
location instead of the development location, so there was no reason
to have made this change in the first place - I had been using tox
incorrectly previously.