Better isolation of LOCALE-based tests.
One of the locale-based tests was generating the `strxfrm` function
outside of a defined locale, but was using it inside the locale. This
created a mismatch in results for users that have a locale set by
default.

To solve, the creation of the `strxfrm` was delayed till inside the
test through the use of a lambda.

This will close issue #67.