Add MacOS support to Travis-CI
This uses a script that was designed to run pyenv in Travis-CI,
so the hope is that this works a bit better. Additionally, this
automatically handles caching of the installed python so subequent jobs
are hopefully more efficient.

ICU is not tested for MacOS because it is known that ICU solves all
the locale issues in a cross-platform way, so it is only necessary to
test against native locale in MacOS.