Add Windows support for Travis-CI
This uses the choco package manager to add python.

pip must be called with "py -m pip" so the "install" step is implemented
specially for windows.

Like MacOS, ICU is not tested for Windows.

There is a bug right now with choco and secret variables, so secrets are
disabled for windows only. This should be fixed ASAP.