Fix path splitter issue on Windows
If a unix-style absolute path is given to the path_splitter on Windows,
the first element was a unix-style path separator and not a Windows path
sparator. This has been fixed.

The docstring has updated to be true for all OSes.

Unit tests are updated to account for this change.