The ns enum actually now is a python enum
I have been wanting to do this for quite some time, but needed the
functionality of the IntEnum for backwards-compatibility. Now that
Python 3.5 is no longer supported, the IntEnum can be used and the
definition of ns is far simpler than it used to be.

This plays nice with mypy also, which is the driving change.