Add tests to demonstrate the PATH ext bug
I'm not sure sure it is *actually* a bug, but the PATH algorithm's way
of splitting extensions was over-zealous and in practice will split off
more extensions that is probably desired. To fix this, we will need to
add a heuristic, but this commit adds tests to demonstrate the problem.